split code and started working on apollo-state-link
