+++ package.json
@@ -16,6 +16,7 @@
     "apollo-client": "^1.2.2",
     "apollo-engine": "^0.5.4",
     "apollo-errors": "^1.4.0",
+    "apollo-link-state": "^0.4.2",
     "apollo-server-express": "^1.2.0",
     "babel-runtime": "^6.26.0",
     "bcrypt": "^2.0.1",
