+++ packages/vulcan-lib/lib/modules/apollo-client/cache.js
@@ -1,0 +1,4 @@
+import { InMemoryCache } from 'apollo-client-inmemory'
+
+const cache = new InMemoryCache()
+export default cache
