+++ packages/vulcan-lib/lib/modules/apollo-client/links/error.js
@@ -1,0 +1,11 @@
+import { onError } from 'apollo-link-error';
+
+const errorLink = onError(({ graphQLErrors, networkError }) => {
+    if (graphQLErrors)
+        graphQLErrors.map(({ message, locations, path }) =>
+            console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`)
+        );
+    if (networkError) console.log(`[Network error]: ${networkError}`);
+});
+
+export default errorLink
