+++ packages/vulcan-lib/lib/modules/apollo-client/links/http.js
@@ -1,0 +1,7 @@
+import { HttpLink } from 'apollo-link-http';
+
+const httpLink = new HttpLink({
+    uri: '/graphql',
+    credentials: 'same-origin',
+});
+export default httpLink
