+++ packages/vulcan-lib/lib/modules/apollo-client/links/meteor.js
@@ -1,0 +1,4 @@
+import { MeteorAccountsLink } from 'meteor/apollo';
+
+const meteorAccountsLink = new MeteorAccountsLink();
+export default meteorAccountsLink
